requirements.txt
aioredis==1.3.1
anyio==4.8.0
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.1.0
autobahn==24.4.2
Automat==24.8.1
boto3==1.35.56
botocore==1.35.56
cachetools==5.5.2
certifi==2024.8.30
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
constantly==23.10.4
cryptography==44.0.0
daphne==4.1.2
defusedxml==0.7.1
distlib==0.3.9
dj-rest-auth==7.0.1
Django==5.2.3
django-allauth==65.7.0
django-cors-headers==4.6.0
django-extensions==3.2.3
django-filter==24.3
django-oauth-toolkit==3.0.1
django-redis==5.4.0
django-storages==1.14.4
djangorestframework==3.16.0
djangorestframework_simplejwt==5.4.0
easy-colors==0.0.1
facebook-sdk==3.1.0
filelock==3.16.1
google-api-core==2.24.1
google-api-python-client==2.162.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.69.0
gunicorn==23.0.0
h11==0.14.0
hiredis==3.1.0
httplib2==0.22.0
httptools==0.6.4
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
jmespath==1.0.1
jwcrypto==1.5.6
msgpack==1.1.0
numpy==2.1.0
oauthlib==3.2.2
packaging==24.2
pillow==11.1.0
platformdirs==4.3.6
proto-plus==1.26.0
protobuf==5.29.3
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==25.0.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-sendmail==0.3.4
python3-openid==3.2.0
PyYAML==6.0.2
redis==5.2.0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
s3transfer==0.10.3
send-mail==0.0.2
service-identity==24.2.0
setuptools==75.8.2
six==1.16.0
sniffio==1.3.1
social-auth-app-django==5.4.3
social-auth-core==4.5.6
sqlparse==0.5.3
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.34.0
virtualenv==20.28.0
watchfiles==1.0.4
websockets==14.2
zope.interface==7.2
